requirements.txt
asgiref==3.7.2
coverage==7.4.3
DateTime==5.4
Django==5.0.2
exceptiongroup==1.2.0
iniconfig==2.0.0
packaging==23.2
pluggy==1.4.0
pytest==8.0.2
pytest-cov==4.1.0
python-dateutil==2.8.2
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
zope.interface==6.2
